SensorService: Add one orientation sensor without fusion support
BZ: 68951

This patch provide an virtual orientation sensor
which is not based on sensor fusion.

Android sensor service provides virtual orientation
sensor based on sensor fusion if the device has
physical accel, magnetic and gyro sensors.
On LEX platform gyro sensor is not available so
virtual orientation sensor is not supported by
sensor service, this cause CTS test failed.
This patch provide an orientation sensor based
on accel and magnetic sensor only to solve this
issue.

Change-Id: I4d6ee8da7816ef7fc5f44020063e61c7dae9ba81
Signed-off-by: jli127 <jian.d.li@intel.com>
Reviewed-on: http://android.intel.com:8080/74791
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Chen, Jason CJ <jason.cj.chen@intel.com>
Reviewed-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Tested-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/76506
Reviewed-by: Yan, Leo <leo.yan@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>