Change the framebuffer format to BGRA base on PVR default setting
BZ: 69058

In MR1, the framebuffer format is hard coded as ARGB8888. But in
IMG's default setting, it can't support HAL RGBA as egl render
surface format. Although we can fix it in PVR code, we will not do
that because it need change many place and that is not a validate
path.

Change-Id: I4b17e3a254bd6b7c82c412cfb5fc4cf9baf8df62
Signed-off-by: Marc Mao <marc.mao@intel.com>
Reviewed-on: http://android.intel.com:8080/74735
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>