Hack FBO flush issue in captureScreen
BZ: 72011

In JB MR1, if hwc HAL version is bigger than 1.1. framebuffer HAL
will be closed by SurfaceFlinger. And SurfaceFlinger will not call
compositionComplete any more. This will cause captureScreen get some
corruption because IMG's OGL driver will not flush FBO context when
switch FBO in middle of a scene. This is a temp hack to this issue.
We add a compositionComplete interface to HWC hal and let it still
call it even HWC version is 1.1 later. If IMG provides a formal
solution for it, we will revert this patch.

Change-Id: I6da80ddcaff70af86dc48a9e27839b359a708967
Signed-off-by: Marc Mao <marc.mao@intel.com>
Reviewed-on: http://android.intel.com:8080/79390
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>