To fix the out-of-memory problem when playing video in VideoEditor
BZ: 74638

Root cause:
1. In JB MR1, the ConsumerBase class is introduced, and it's used as
the base class for SurfaceTexture.

2. There is a virtual function called freeLockedBuffer in both
ConsumerBase and SurfaceTexture.

3. When a SurfaceTexture instance is deconstructed, the ConsumerBase
deconstrutor is called. In ~ConsumerBase, freeLockedBuffer() is called.
The author's intent is to call the freeLockedBuffer() in the derived
SurfaceTexture class. However, according to the C++ language,
only the freeLockedBuffer() in ConsumerBase is called.

4. This leads to wrong GraphicBuffer reference counts. When a
SurfaceTexture instance is destroyed in video editor, the GraphicBuffer
decontructor is not called. The allocated GraphicBuffer instances are not
freed correctly, which causes the sharedbuffer in the Gralloc HAL
out-of-memory...

The solution is that use the SurfaceTexture decontructor to all
the SurfaceTexture::freeLockedBuffer().

Change-Id: Ia2477100254d28ca2f63aa09651e8eee8b601307
Signed-off-by: wfeng6 <wei.feng@intel.com>
Reviewed-on: http://android.intel.com:8080/85099
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: Wang, Yi A <yi.a.wang@intel.com>
Reviewed-by: Cheng, Yao <yao.cheng@intel.com>
Reviewed-by: Gu, Wangyi <wangyi.gu@intel.com>
Reviewed-by: Jiang, Fei <fei.jiang@intel.com>
Reviewed-by: Liang, Dan <dan.liang@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>