SurfaceFlinger: not bypass composition on HDMI when screen on
BZ: 77275

If turning off screen in landscape mode and later turning on screen,
composition on HDMI is bypassed, surfaceflinger is waiting for the
rotation animation to start, then it will start composition on HDMI
display. This causes the lock screen not shown on HDMI display.

Fixed by not bypassing composition on external display when screen is
just acquired.

Change-Id: If2312877534f96a37d6e3c4cefe7af6f81dd35ca
Signed-off-by: Hong Liu <hong.liu@intel.com>
Reviewed-on: http://android.intel.com:8080/85160
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>