SurfaceFlinger: fix random 90 degree image on HDMI during rotation
BZ: 84763

DO NOT UPSTREAM!

this patch fixes the random 90 degree image on HDMI during rotation.
It is due to the mDisplay Vector order exchange sometimes when boot
up. Display device may be EXTERNAL DISPLAY when dpy=0. So here use
getDisplayType to check whether it is PRIMARY_DISPLAY instead of dpy.

Change-Id: Id796616217c80cb73866f656efd524b9fd34151f
Signed-off-by: Lei Zhang <lei.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/90002
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>