[PORT FROM MAIN] Video-HDMI: Increase Buffer queue size to 2 to avoid buffer overwrite issue and improve video playback quality
BZ: 90548

1: There is a timing sequence issue between video queue buffer and
SurfaceFlinger's vsync. Video render is different with GUI, GUI will
wait a vsync from HWC, then queue buffer, but video is not.If video
queue buffer is not smooth or missed a vsync, video buffer will be over
written. So increase buffer's number to 2 to avoid buffer overwrite,
and 1 video latency is acceptable for A/V sync;
2: This patch has a confict with the patch 90981 - 'Vsync mechnism
enhancement', so I abandoned this patch, but, the patch 90981 caused a
serious regression 93506 and had been reverted,
you can refer the patch 97474, so I restores this patch;
3: This patch is only for video playback quality;

Change-Id: Ie24ff052f934d5ca7e7afda7c4a2608e90d75df2
Orig-Change-Id: I0a80ddcb70f78f44b6d93d5437e6d20c6db201ea
Signed-off-by: Zhu,Tianyang <tianyang.zhu@intel.com>
Signed-off-by: Lin Xie <lin.xie@intel.com>
Reviewed-on: http://android.intel.com:8080/122113
Reviewed-by: Liang, Dan <dan.liang@intel.com>
Reviewed-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>