Fix segv in surfaceflinger when pressing back key while playing a video.
BZ: 91419

When the back key is pressed while playing a video, the methods
BufferQueue::consumerDisconnect and BufferQueue::disconnect are both called.
These are called by the consumer and producer for the buffers being managed.
If a sleep of about a half second is inserted at the beginning of disconnect,
the problem doesn't occur. What is happening is that consumerDisconnect gets
a chance to run before disconnect, and consumerDisconnect completely removes
all buffers in the queue. Without the sleep, disconnect will sometimes run
before consumerDisconnect, leaving the buffer head still in the queue, and a
SurfaceFlinger thread will try to use the SurfaceTexture module to render
using that buffer. This results in the method SurfaceTexture::createImage
being called with a NULL value for graphicBuffer.  createImage did not check
the value of graphicBuffer, or the value returned by graphics->getNativeBuffer()
for NULL, resulting in a bad third paramenter being passed to eglCreateImageKHR,
resulting in a segmentation violation. These checks have been added in this
change.

Change-Id: I2fb41badf2b9928a2113199859469ae84877f7b0
Signed-off-by: David Pinedo <davidx.pinedo@intel.com>
Reviewed-on: http://android.intel.com:8080/97391
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Dai, Yu <yu.dai@intel.com>
Reviewed-by: Stimson, Dale B <dale.b.stimson@intel.com>
Tested-by: Stimson, Dale B <dale.b.stimson@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>