widi: Hardcode HWC display id to 2 for virtual displays
BZ: 94285

DO NOT UPSTREAM

We need widi's screen contents to be sent to HWC for widi's extended
video mode to work properly. Unfortunately, that only happens on the
first connection.

hwcomposer.h says the following about HWC 1.2:

> For HWC 1.2 and later, numDisplays will be HWC_NUM_DISPLAY_TYPES or more.
> The extra entries correspond to enabled virtual displays, and will be
> non-NULL. In HWC 1.2, support for one virtual display is required, and
> no more than one will be used. Future HWC versions might require more.

SurfaceFlinger has an incomplete implementation for sending multiple
virtual displays to HWC. There is dynamic assignment for virtual
displays ids via HWComposer::allocateDisplayId(), but since id 0 and 1
are reserved (for the primary and HDMI displays) and mNumDisplays is 3
(assuming HWC 1.2), a virtual display can only be assigned an HWC id
of 2.

After a display has received an id of 2, no more displays can get that
id even after disconnection because HWComposer::freeDisplayId() is
never called. Therefore, SF will handle the new displays itself
without involving HWC. If we did add a call to freeDisplayId(), we
would still have a problem if you disconnect and reconnect in a short
time interval because SF doesn't remove the display until the
DisplayToken held at the Java layer is garbage collected.

So my solution is to hardcode the id for a virtual display to 2, just
like the id is always 0 for primary and 1 for external (HDMI). It is
possible to have more than one active virtual display, but only
through hidden API or a hidden developer option (simulate secondary
displays). The only user-visible way to create a virtual display is
widi, which limits you to one at a time.

If multiple displays do have the same id (because of simulate
secondary displays), the last display the loop iterates over is the
one that determines the layer list send to HWC. All displays are still
rendered.

In addition to the above issue, disconnectDisplay() was being ignored
for virtual displays, meaning we kept sending the stale layer list to
HWC after disconnecting widi, so I fixed that as well.

Change-Id: I295e9d96f9a48f1af392e1f8d167a5e8944a5d42
Signed-off-by: Brian Rogers <brian.e.rogers@intel.com>
Reviewed-on: http://android.intel.com:8080/97833
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Veeramani, Karthik <karthik.veeramani@intel.com>
Reviewed-by: Liu, Hong <hong.liu@intel.com>
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Reviewed-by: Chowdhury, Rajneesh <rajneesh.chowdhury@intel.com>
Tested-by: Post, DavidX J <davidx.j.post@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>