[PORT FROM MAIN] To fix the out-of-memory problem when playing video in VideoEditor
BZ: 74638

Root cause:
1. In JB MR1, the ConsumerBase class is introduced, and it's used as
the base class for SurfaceTexture.

2. There is a virtual function called freeLockedBuffer in both
ConsumerBase and SurfaceTexture.

3. When a SurfaceTexture instance is deconstructed, the ConsumerBase
deconstrutor is called. In ~ConsumerBase, freeLockedBuffer() is called.
The author's intent is to call the freeLockedBuffer() in the derived
SurfaceTexture class. However, according to the C++ language,
only the freeLockedBuffer() in ConsumerBase is called.

4. This leads to wrong GraphicBuffer reference counts. When a
SurfaceTexture instance is destroyed in video editor, the GraphicBuffer
decontructor is not called. The allocated GraphicBuffer instances are not
freed correctly, which causes the sharedbuffer in the Gralloc HAL
out-of-memory...

The solution is that use the SurfaceTexture decontructor to all
the SurfaceTexture::freeLockedBuffer().

Category:aosp improvement
Domain: Video.Media-video editor
Origin: internal
Upstream-Candidate: yes

Change-Id: Ic9a602e617f62a5288687f413129200d30b4ca2e
Orig-Change-Id: Ibc7be92a0824248e0af1283f46a80f3116f89e53
Signed-off-by: wfeng6 <wei.feng@intel.com>
Signed-off-by: Dan Liang <dan.liang@intel.com>
Signed-off-by: Yuanjun Huang <yuanjun.huang@intel.com>