[PORT FROM R42]Workaround game and map flicking issues on clv+ platform.
BZ: 122150

The flicking issue is not easy to debug directly because any action to
capture the buffer will slow down the system and the flicking
disappeared. So we traced to the root cause via lots of bisect
experiments.

Surfaceflinger's EGL context priority is set to high in BZ:74200 and
it's to tune performance on mfld platforms. So set it back to default
(medium) on clv+ platform.

Change-Id: Icbb5b33341ace91102069a3b7f2c66513131deb1
Signed-off-by: Xinyun Liu <xinyun.liu@intel.com>
Signed-off-by: Thomas Buhot <thomas.bohot@intel.com>
Signed-off-by: Stanley Lee <stanley.lee@intel.com>
Signed-off-by: Randy Xu <randy.xu@intel.com>
Signed-off-by: Yun Tu(Mark Tu) <yun.tu@intel.com>
Signed-off-by: yifeix.xue <yifeix.xue@intel.com>
Reviewed-on: http://android.intel.com:8080/118127
Reviewed-by: Liu, MinX Z <minx.z.liu@intel.com>
Tested-by: Liu, MinX Z <minx.z.liu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>