[PORT FROM R42B-STABLE] SurfaceFlinger: 18 bit post process dithering
BZ: 137776

This checkin enables dithering as a post processing pass after the
main composition has completed. Its enabled only when the
appropriate make target is set for building with an 18 bit panel.

This is a quality improvement (at a power cost) from the previous
solution of using a 16 bit render target. It performs half of
the dithering in the 32 bit color space by adding on an offset in
a 4x4 pattern. The expectation is that the display pipe will
later truncate and the offset causes the dithering effect to
occur on the screen.

The issue with the previous solution is that we would loose the
alpha channel on dithering to 565. This means we cannot use any
sprite planes, and hence means that we cannot support protected
video. By retaining the alpha channel, this commit allows us
to use the sprite plane for all video.

The dithering is only applied to a Surfaceflinger render target as
requested by the Hardware Composer.

Change-Id: I48ca14aa223b8e3626fb6be12fc9220692e9ab62
Signed-off-by: Abel Bernabeu <abel.bernabeu@intel.com>
Reviewed-on: http://android.intel.com:8080/132920
Reviewed-by: Sun, Hang L <hang.l.sun@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>