[PORT FROM MAIN] widi : Hardcode HWC display id to 2 for virtual displays.
BZ: 94285

DO NOT UPSTREAM
We need widi's screen contents to be sent to HWC for widi's extended
video mode to work properly. Unfortunately, that only happens on the
first connection.

SurfaceFlinger has an incomplete implementation for sending multiple
virtual displays to HWC. There is dynamic assignment for virtual
displays ids via HWComposer::allocateDisplayId(), but since id 0 and 1
are reserved (for the primary and HDMI displays) and mNumDisplays is 3
(assuming HWC 1.2), a virtual display can only be assigned an HWC id
of 2.
After a display has received an id of 2, no more displays can get that
id even after disconnection because HWComposer::freeDisplayId() is
never called. Therefore, SF will handle the new displays itself
without involving HWC. If we did add a call to freeDisplayId(), we
would still have a problem if you disconnect and reconnect in a short
time interval because SF doesn't remove the display until the
DisplayToken held at the Java layer is garbage collected.

So my solution is to hardcode the id for a virtual display to 2, just
like the id is always 0 for primary and 1 for external(HDMI). It is
possible to have more than one active virtual display. but only
through hidden API or a hidden developer option (simulate secondary
displays). The only user-visible way to create a virtual display is
widi, which limits you to one at a time.

If multiple displays do have the same id (because of simulate
secondary displays), the last display the loop iterates over is the
one that determines the layer list send to HWC. All displays are still
rendered.

we support hwc 1.2 version in mr1 so same has been corrected as well.

Change-Id: I5a42c65c116bde6b19e2d42b9ee922be691586f0
Orig-Change-Id: I295e9d96f9a48f1af392e1f8d167a5e8944a5d42
Signed-off-by: Ashish Singhi <ashish.singhi@intel.com>
Reviewed-on: http://android.intel.com:8080/124595
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>