Integrate Arkham to mainline
BZ: 127460

Arkham is a deeply integrated security container capability for
Android for the purpose of isolation and management in mixed used
cases. Primarily Arkham is intended to enable separation of work and
personal content and capability.

Detailed information about the Arkham requirements and architecture can
be found here: https://sharepoint.amr.ith.intel.com/sites/SDI/Arkham/

This patch modifies the SurfaceFlinger service to deny capturing the screen
while container applications are running. We rely on modifications in the
ActivityManager service to use a system property variable. Whenever a container
application is switched to and cleared when we switch to a non-container
application.

Category: feature differentiation
Domain: AOSP.Extend-Arkham
Origin: Internal
Upstream-Candidate: No, proprietary

Change-Id: I1ecc034c3f64ebc5c21d3fe3008f3346b13021c2
Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>
Reviewed-on: http://android.intel.com:8080/126920
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>