[PORT FROM MAIN] WIDI: Widi is not able to reconnect,after the first connection..
BZ: 131842

The root cause is even though the Display ID of the virtual device
is hardcoded to 2 (value of DISPLAY_VIRTUAL), while disconnecting,
it was getting freed up. This should not be done,the virtual
display id is pre-allocated, when surface flinger comes up.

Hence ensured that the display id 2 is not freed, as is done
for 0 and 1, upon disconnecting the virtual device.

Change-Id: Id6062af901e16f2443f53dcf516ade1dec4a20d3
Signed-off-by: Arindam Roy <arindamx.roy@intel.com>