[PORT FROM MAIN] widi: Changes needed in servicemanager to register widi service
BZ: 125777

Widi needs to be registered as a service with servicemanager so that
it can be discovered by other processes at the native layer to talk to
widi service. This patch adds this capability, and uses INTEL_WIDI macro
to conditionally compile it in if the flag is passed during compilation.

This widi service is used by sigma agent and will be included only in userdebug
and eng build variants.

Change-Id: I2721c7f32bfa3f598a8f6b4833e0fe68b1b183af
Orig-Change-Id: I08961aae39db6d5efccaab1757acedd92b8f2cfc
Category: feature differentiation
Domain: WIDI
Origin: internal
Upstream-Candidate: no
Signed-off-by: mamatha balguri <mamatha.balguri@intel.com>
Signed-off-by: Arindam Roy <arindamx.roy@intel.com>
Signed-off-by: Padath Peedikayil Abdul Rahim, AslamX <aslamX.padath.peedikayil.abdul.rahim@intel.com>
Reviewed-on: http://android.intel.com:8080/128652
Reviewed-by: Veeramani, Karthik <karthik.veeramani@intel.com>
Reviewed-by: Padath Peedikayil Abdul Rahim, AslamX <aslamx.padath.peedikayil.abdul.rahim@intel.com>
Reviewed-by: Sun, Hang L <hang.l.sun@intel.com>
Tested-by: Post, DavidX J <davidx.j.post@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
Signed-off-by: Arindam Roy <arindamx.roy@intel.com>