Fix for eglCreateContext error handling
BZ: 144421

In eglCreateContext, in case of an error in underlying egl library,
the error is read by the wrapper, but is not later set again.
As a result, error message will be unavailable to the user.
Patch resets this error message. Problem is visible since
Android 4.3 in GTF3 conformance tests for EGL_KHR_create_context.
This extension must be enabled to allow proper support for ES 3.0
context in Android Apps.

Change-Id: I2d986380cc7fc168026a7b9a1778dbda40071134
Signed-off-by: jbyczkow <jakub.byczkowski@intel.com>