Integrate Arkham to mainline
BZ: 156819

Arkham is a deeply integrated security container capability for
Android for the purpose of isolation and management in mixed used
cases. Primarily Arkham is intended to enable separation of work and
personal content and capability.

Detailed information about the Arkham requirements and architecture can
be found here: https://sharepoint.amr.ith.intel.com/sites/SDI/Arkham/

This patch modifies the SurfaceFlinger service to deny capturing the screen
while container applications are running. We rely on modifications in the
ActivityManager service to use a system property variable. Whenever a container
application is switched to and cleared when we switch to a non-container
application.

Change-Id: I46c97a2476265cbea8b65c3722199de079e30d3a
Category: feature differentiation
Domain: AOSP.Security-Arkham
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>
Reviewed-on: http://android.intel.com:8080/126920
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: cactus <cactus@intel.com>