Enable Virtual Orientation Sensor if there is not one.
BZ: 296

Enable Virtual Orientation Sensor if there is not one. Android L will
disable virtual sensors if they are already available in HAL and replace the
Orientation Sensor from HAL if the virtual rotation vector sensor is in using.
But if there is no rotation vector sensor in HAL, the method replaceAt() will fail.

Change-Id: Ibb2059a3fc5eecfcc66149dd996c389dd5780dfe
Signed-off-by: Han, He <he.han@intel.com>
Signed-off-by: Jiantao Zhou <jiantao.zhou@intel.com>