Fix meta_data_event index error in KeyedVector in SensorService.
BZ: 188628

SensorService record last sensor data event for non-continous sensor, and this
is using KeyedVector and use 'handle' as the key for each sensor. But meta_data_event
is a special event which shows flush success finished and not belong to any sensor,
so its 'handle' is set to zero(0) due the requirement, and the key zero is not in the
KeyedVector, so it will cause crash when try to do some action using this key.
And also meta_data_event should not recorded.

Change-Id: Ib43d2a2e16a9f8d12e642fa00ce94d2411bf02cf
Category: aosp improvement
Domain: SENSORS.VIRT-Common
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Han, He <he.han@intel.com>