Drop the oldest buffer when queued buffer size is larger than 2
BZ: 157352
When the decoder queue buffer timing is in asynchronous with the vsync, more
than 2 buffers managed by the BufferQueue will be in the QUEUED status.
The dequeuebuffer call will take a long time (10ms for example). This leads
to frame drops in A/V sync phase of the AwesomePlayer. In this patch, the
queued buffer size is >2, the oldest buffer will be dropped. So dequeueBuffer
will NOT be blocked, and the frame drops in the A/V sync phase is reduced
significantly.

Change-Id: I2cecf95d6807d51f6ae258b41b6756efe318b705
Category: aosp improvement
Domain: Video.Media-framework
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: wfeng6 <wei.feng@intel.com>
Signed-off-by: ywan171 <yi.a.wang@intel.com>