Fix VirtualDisplaySurface initialisation bugs
BZ: 166685
BZ: 167776

A recent Google commit in mr1 changed the way the VirtualDisplaySurface was
initialised. I suspect they intended to not reset the fence whenever
resetPerFrameState was called, but this function was also used to
initialise the mFbFence and mProducerBuffers members.

As a result of this, when the Gralloc suceeds in allocating a
HAL_PIXEL_FORMAT_IMPLEMENTATION_DEFINED buffer, and the HWC chooses
to completely compose the first frame of the virtual display,
SurfaceFlinger crashes dereferencing a null pointer.

The commit that introduced this bug was:
https://android.googlesource.com/platform/frameworks/native/+/14e8b01a761180250671f3d6494532d8bd04fa8e%5E%21

Change-Id: I6fd9c1c8d13bcd13b436b29d5570b45c6e1e14dc
Category: aosp improvement
Domain: Graphics-SF
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Gary Smith <gary.k.smith@intel.com>
Reviewed-on: https://android.intel.com/162017
Tested-by: sys_abtbuild <sys_abtbuild@intel.com>
Reviewed-by: Chiras, Robert <robert.chiras@intel.com>
Reviewed-by: Serban, Mihai <mihai.serban@intel.com>